Rename ApplicationOptions to ApplicationProperties
