+++ src/main/java/edu/project2/solving/dfs/DFSSolver.java
@@ -3,7 +3,7 @@ package edu.project2.solving.dfs;
 import edu.project2.model.Cell;
 import edu.project2.model.Maze;
 import edu.project2.model.Node;
-import edu.project2.options.ApplicationOptions;
+import edu.project2.properties.ApplicationProperties;
 import edu.project2.solving.Solver;
 import java.util.ArrayDeque;
 import java.util.Collections;
@@ -68,7 +68,7 @@ public class DFSSolver implements Solver {
     }
 
     private Optional<Node> getFirstNeighbor(Node cur) {
-        for (int[] delta : ApplicationOptions.DELTAS) {
+        for (int[] delta : ApplicationProperties.DELTAS) {
             int row = cur.getRow() + delta[0];
             int column = cur.getColumn() + delta[1];
