+++ src/main/java/edu/project2/view/console/ConsoleRenderer.java
@@ -2,7 +2,7 @@ package edu.project2.view.console;
 
 import edu.project2.model.Cell;
 import edu.project2.model.Maze;
-import edu.project2.options.ApplicationOptions;
+import edu.project2.properties.ApplicationProperties;
 import edu.project2.view.Renderer;
 import java.util.List;
 
@@ -24,11 +24,11 @@ public class ConsoleRenderer implements Renderer {
         for (Cell[] row : maze.getGrid()) {
             for (Cell cell : row) {
                 if (cell.isWall()) {
-                    sb.append(ApplicationOptions.WALL_STRING);
+                    sb.append(ApplicationProperties.WALL_STRING);
                 } else if (showEscape && cell.isEscape()) {
-                    sb.append(ApplicationOptions.PATH_STRING);
+                    sb.append(ApplicationProperties.PATH_STRING);
                 } else {
-                    sb.append(ApplicationOptions.PASSAGE_STRING);
+                    sb.append(ApplicationProperties.PASSAGE_STRING);
                 }
             }
             sb.append('\n');
